feat Manage read-only lines in a unit test file.
It makes no sense to modify some of the lines in the unit test file template.
These lines are vital for the unit test file to run properly. In order to keep the
developer from unintentional modifications these lines have been made
read-only. No editions are allowed. The lines get highlighted when encountered.