refactor Highlight multiple read-only lines
Only one read-only line can be highlighted at a time. This is not ideal
as the text selection may include more than one read-only line. The
high complexity of highlighting and preventing the edition of a
read-only means these features will exist in isolation. The incomplete
implementation of the latter has been removed for now.