refactor: Remove event constructors polyfill
Before, the unit test file template had a script tag that loaded the event constructors polyfill. That was not necessary in all cases. For example, certain unit test files had nothing to do with an Internet browser. Those unit tests could also well be run in a non-Internet-browser environment. For that reason, the event constructors polyfill is now removed. If needed for a unit test file, it will have to be added manually.